+++ src/steps/vmdr/converters.test.ts
@@ -626,6 +626,9 @@ describe('getHostDetails', () => {
     expect(getHostDetails({ fqdn: 'THIS.IS.MY.HOST' })).toMatchObject({
       fqdn: 'this.is.my.host',
     });
+    expect(getHostDetails({ fqdn: {} as any })).toMatchObject({
+      fqdn: undefined,
+    });
     expect(getHostDetails({ fqdn: undefined })).toMatchObject({
       fqdn: undefined,
     });