+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+### Fixed
+
+- #101 Fixed error handling host details with
+  `host.fqdn.toLowerCase not a function`
+
 ## 5.4.2 - 2021-04-20
 
 ### Fixed